requirements.txt
telethon>=1.28.0,<2.0
python-dotenv>=1.0.0,<2.0
Pillow>=10.0.0,<11.0
